[Log View] "Copy Logtext to clipboard" icon is no longer supported for TBs.
The "Copy Logtext to clipboard" icon is no longer supported for TBs.

The icon is not always available for TB Logs.
Background: After adding a TB log and after changing a TB log, you will automatically be taken to the View Log page. The View Log page is not accessed, only the page content is exchanged. Therefore the logtext is not available on the View Log page.

Clarify the folowing issues:
- [Log View] "Copy Logtext to clipboard" icon not available after post a TB log. #2513 
- [Log Edit, Log View] "Copy Logtext to clipboard" icon does not contain the changed TB logtext after an update. #2523

Further information are in pull request #2522

---
close #2513
close #2523