[Log Form] Fix issue in FF
If FF loads the page from browser cache then it is already fully loaded when the observing starts and thus the observer callback isn't called.
To circumvent this, we add a dummy element (and immediately remove it) to force a call to the observer callback.
Note that in Chrome or if one forces a page reload from the server this isn't an issue.