[Config] Keep user defined order of map layers
One can reorder the map layers in config, but after config close and reopen, the map layers are ordered alphabetically - even if the corresponding option is not active.
Now the user defined order is preserved.

Moreover, the layer order is preserved in maps - provided the option 'Sort map layers in map' is not active.